+++ src/main/java/de/adorsys/keycloak/config/repository/ClientRepository.java
@@ -183,6 +183,13 @@ public class ClientRepository {
         try (Response response = clientResource.authorization().resources().create(resource)) {
             CreatedResponseUtil.getCreatedId(response);
         } catch (WebApplicationException e) {
+            if (e.getResponse().getStatus() == HTTP_NOT_FOUND || e.getResponse().getStatus() == HTTP_NOT_IMPLEMENTED) {
+                throw new KeycloakRepositoryException(
+                        String.format("Authorization API not supported for client '%s' in realm '%s' (likely FGAP V2 active)",
+                                clientResource.toRepresentation().getClientId(), realmName), e);
+            }
+            throw e;
+        } catch (WebApplicationException e) {
             handleAuthorizationApiException(e, clientResource, realmName);
         }
     }
@@ -223,6 +230,13 @@ public class ClientRepository {
         try (Response response = clientResource.authorization().scopes().create(scope)) {
             CreatedResponseUtil.getCreatedId(response);
         } catch (WebApplicationException e) {
+            if (e.getResponse().getStatus() == HTTP_NOT_FOUND || e.getResponse().getStatus() == HTTP_NOT_IMPLEMENTED) {
+                throw new KeycloakRepositoryException(
+                        String.format("Authorization API not supported for client '%s' in realm '%s' (likely FGAP V2 active)",
+                                clientResource.toRepresentation().getClientId(), realmName), e);
+            }
+            throw e;
+        } catch (WebApplicationException e) {
             handleAuthorizationApiException(e, clientResource, realmName);
         }
     }
@@ -254,6 +268,13 @@ public class ClientRepository {
 
         try (Response response = clientResource.authorization().policies().create(policy)) {
             CreatedResponseUtil.getCreatedId(response);
+        } catch (WebApplicationException e) {
+            if (e.getResponse().getStatus() == HTTP_NOT_FOUND || e.getResponse().getStatus() == HTTP_NOT_IMPLEMENTED) {
+                throw new KeycloakRepositoryException(
+                        String.format("Authorization API not supported for client '%s' in realm '%s' (likely FGAP V2 active)",
+                                clientResource.toRepresentation().getClientId(), realmName), e);
+            }
+            throw e;
         } catch (WebApplicationException e) {
             handleAuthorizationApiException(e, clientResource, realmName);
         }