+++ src/test/java/de/adorsys/keycloak/config/service/ImportClientsIT.java
@@ -1570,6 +1570,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(41)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldAddAuthzPoliciesForRealmManagement() throws IOException {
         doImport("41_update_realm_add_authz_policy_realm-management.json");
@@ -1692,6 +1693,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(42)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldUpdateAuthzPoliciesForRealmManagement() throws IOException {
         doImport("42_update_realm_update_authz_policy_realm-management.json");
@@ -1842,6 +1844,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(43)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldRemoveClientAndAuthzPoliciesForRealmManagement() throws IOException {
         doImport("43_update_realm_remove_client_and_authz_policy_realm-management.json");
@@ -1953,6 +1956,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(44)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldRemoveAuthzPoliciesForRealmManagement() throws IOException {
         doImport("44_update_realm_remove_authz_policy_realm-management.json");
@@ -2039,6 +2043,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(45)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldUpdateAuthzPoliciesPerIdentityProvidersForRealmManagement() throws IOException {
         doImport("45_update_realm_update_authz_policy_for_idp_realm-management.json");
@@ -2106,6 +2111,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(46)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldUpdateAuthzPoliciesPerIdentityProvidersWithPlaceholdersForRealmManagement() throws IOException {
         doImport("46_update_realm_update_authz_policy_for_idp_with_placeholder_realm-management.json");
@@ -2179,6 +2185,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(47)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldUpdateAuthzPoliciesPerRolesWithPlaceholdersForRealmManagement() throws IOException {
         doImport("47_update_realm_update_authz_policy_for_role_with_placeholder_realm-management.json");
@@ -2250,6 +2257,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(48)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldUpdateAuthzPoliciesPerGroupsWithPlaceholdersForRealmManagement() throws IOException {
         doImport("48_update_realm_update_authz_policy_for_group_with_placeholder_realm-management.json");
@@ -2355,6 +2363,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(50)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void shouldNotTriggerErrorWhenReferencingInvalidUuidInFineGrainedAuthz() throws IOException {
         // These scenarios do not use placeholders and instead reference objects by UUID - which do not need to exist.
@@ -2380,6 +2389,7 @@ class ImportClientsIT extends AbstractImportIT {
 
     @Test
     @Order(51)
+    @DisabledIfSystemProperty(named = "keycloak.version", matches = "26\\.[2-9].*", disabledReason = "FGAP V1 APIs removed in Keycloak 26.2.0+, use FGAP V2 compatibility tests instead")
     @DisabledIfSystemProperty(named = "keycloak.version", matches = "2[6-9]\\.[2-9].*|[3-9][0-9].*", disabledReason = "FGAP V2 active in KC 26.2+, realm-management authorization handled differently")
     void updateRealmWithClientWithMoreThan100RolesInRealmManagementAuthorization() throws IOException {
         doImport("51_update_realm_with_client_with_more_than_100_roles_in_realm_management_authorization.json");